requirements.txt
reportlab==4.2.0
Faker==25.8.0
pypdf2==3.0.1
pikepdf==9.0.0
streamlit==1.36.0
stqdm==0.0.5
PyMuPDF==1.24.7 
pillow==10.4.0
pytesseract==0.3.10 
markdownify==0.12.1
openai-whisper==20231117
streamlit-audiorecorder==0.0.5
plotly==5.22.0